+++ packages/observable-client/tests/mockSubstrateClient.ts
@@ -1,4 +1,3 @@
-import { BlockInfo } from "@/chainHead"
 import {
   ChainHead,
   FollowEventWithRuntime,