+++ packages/client/CHANGELOG.md
@@ -23,7 +23,7 @@
 - BREAKING: When creating transactions the optional `at` value only accepts specific block-hashes.
 - BREAKING: `getBody` and `getHeader` only accept block-hashes.
 - BREAKING: renamed `watchBlockBody` to `getBlockBody$`
-- BREAKING: `watchValue` emits the block the value was found in, and whether the value is finalized.
+- BREAKING: `watchValue` emits the block the value was found in.
 - BREAKING: `watchValue` options are now consistent with `watchEntries` options, and also with the - BREAKING: rename chainspecs `ksmcc3` => `kusama`, `westend2` => `westend`
 
 - **WS-Provider:**